% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decrypt_instances.R
\name{decryptInstances}
\alias{decryptInstances}
\title{Decrypt encrypted instance files}
\usage{
decryptInstances(instanceRegistry, privateRSAKey, writeToPath)
}
\arguments{
\item{instanceRegistry}{Instance registry dataframe generated by getEncryptedInstances}

\item{privateRSAKey}{Private RSA key (base64 encoded)}

\item{writeToPath}{Path to directory where to save decrypted instance files}
}
\value{
Instance registry dataframe with paths to individual decrypted instances files
}
\description{
Decrypt encrypted instance files
}
