% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_encrypted_instances.R
\name{getEncryptedInstances}
\alias{getEncryptedInstances}
\title{Retrieve encrypted instance files of a form}
\usage{
getEncryptedInstances(
  instanceRegistry,
  writeToPath,
  baseDownloadUrl = "https://api.ona.io"
)
}
\arguments{
\item{instanceRegistry}{Instance registry dataframe generated by getInstanceRegistry}

\item{writeToPath}{Path to directory where to save encrypted instance files}

\item{baseDownloadUrl}{Optional, base download url from API}
}
\value{
Instance registry dataframe with paths to individual encrypted instances files
}
\description{
Retrieve encrypted instance files of a form
}
