% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forms.R
\name{getForms}
\alias{getForms}
\title{Get information on all active forms from API}
\usage{
getForms(token, baseUrl = "https://api.ona.io/api/v1/")
}
\arguments{
\item{token}{Identification token to connect to API}

\item{baseUrl}{Optional, base url of API}
}
\value{
Dataframe with updated information on active forms
}
\description{
Get information on all active forms from API
}
