% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instance_registry.R
\name{getInstanceRegistry}
\alias{getInstanceRegistry}
\title{Get a list of all instances files submitted to a form}
\usage{
getInstanceRegistry(xFormId, token, baseUrl = "https://api.ona.io/api/v1/")
}
\arguments{
\item{xFormId}{xForm ID of the form}

\item{token}{Identification token to connect to API}

\item{baseUrl}{Optional, base url of API}
}
\value{
Instance registry dataframe with all instances submitted to the form
}
\description{
Get a list of all instances files submitted to a form
}
