% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_form_directory.R
\name{setupFormDirectory}
\alias{setupFormDirectory}
\title{Setup directory to retrieve and aggregate data of a form}
\usage{
setupFormDirectory(
  xFormId,
  token,
  formDirectory = NULL,
  baseUrl = "https://api.ona.io/api/v1/"
)
}
\arguments{
\item{xFormId}{xForm ID of the form}

\item{token}{Identification token to connect to API}

\item{formDirectory}{Optional, path to directory where to save form information and data}

\item{baseUrl}{Optional, base url of API}
}
\value{
Form info object containing information on the form and its directory
}
\description{
Setup directory to retrieve and aggregate data of a form
}
